Refactored the task table
